Add: handle mixed inputs
- If given a list of files that list should be returned so if a function
can take in either a list of files or a directory it can pass the input
into LISTFILES without having to check which it is and be assured it
will get a list of files back. Should also be able to handle a mix of
files and directories.
- Checks each element in the input cell to see if it's a file,
directory, glob pattern, or regex pattern and list files for each
element based on the result.
- Functions can now pass any input into LISTFILES, whether a list of
directories, a single file, or any other type, and always get a list of
files back.